to sleep

+++ gauss_cuda.cu
@@ -45,12 +45,12 @@ blurEffect(double *d_kernel, int height, int width,  unsigned char *d_R,  unsign
             double blueTemp = 0;
             double greenTemp = 0;
             double acum = 0;
-            for (int k = 0; k < (int)kernelSize; k++ )
+            for (int k = 0; k < kernelSize; k++ )
             {
                 int y = (i - radius + k + height)%height;
                 for (int l = 0; l < kernelSize; l++)
                 {
-                    int x = (j - radius + l + width )%width;
+                    int x = (j - radius + l + width )% width;
                     // x = x < 0 ? 0 : x < width ? x : width - 1;
                     redTemp += d_R[y*width + x] * d_kernel[k*kernelSize + l];
                     greenTemp += d_G[y*width + x] * d_kernel[k*kernelSize + l];
